requirements.txt
streamlit==1.31.0
google-api-python-client==2.118.0
beautifulsoup4==4.12.3
requests==2.31.0
sentence-transformers==2.5.1
google-generativeai==0.3.2
python-dotenv==1.0.1
aiohttp==3.9.3
numpy==1.26.4
pandas==2.2.1
tqdm==4.66.2
